feat: add summary citation to discover API + DP API
