+++ backend/cellguide/pipeline/source_collections/source_collections_generator.py
@@ -1,10 +1,10 @@
-from backend.cellguide.pipeline.canonical_marker_genes.utils import format_citation_dp
 from backend.cellguide.pipeline.source_collections.types import SourceCollectionsData
 from backend.common.census_cube.utils import (
     descendants,
     get_collections_from_discover_api,
     get_datasets_from_discover_api,
 )
+from backend.common.citation import format_citation_dp
 
 
 def generate_source_collections_data(all_cell_type_ids_in_corpus: list[str]) -> dict[str, list[SourceCollectionsData]]:
