+++ backend/curation/api/v1/curation/collections/common.py
@@ -3,6 +3,7 @@ from typing import List, Optional, Tuple, Union
 from urllib.parse import urlparse
 from uuid import UUID
 
+from backend.common.citation import format_citation_dp
 from backend.common.corpora_config import CorporaConfig
 from backend.common.utils.http_exceptions import (
     ForbiddenHTTPException,
@@ -294,6 +295,7 @@ def reshape_dataset_for_curation_datasets_index_api(
     ds["collection_name"] = collection_version.metadata.name
     doi, _ = extract_doi_from_links(collection_version.metadata.links)
     ds["collection_doi"] = doi
+    ds["collection_doi_label"] = format_citation_dp(collection_version.publisher_metadata)
 
     # At this point, dataset shape for public requests are complete.
     if is_dataset_visibility_public:
