+++ tests/unit/backend/common/test_crossref_provider.py
@@ -1,12 +1,12 @@
 import copy
 import json
 import unittest
-from unittest.mock import patch
+from unittest.mock import Mock, patch
 
 from requests import RequestException
 from requests.models import HTTPError, Response
 
-from backend.layers.thirdparty.crossref_provider import (
+from backend.layers.common.provider.crossref_provider import (
     CrossrefDOINotFoundException,
     CrossrefException,
     CrossrefFetchException,
@@ -16,15 +16,15 @@ from backend.layers.thirdparty.crossref_provider import (
 
 
 class TestCrossrefProvider(unittest.TestCase):
-    @patch("backend.layers.thirdparty.crossref_provider.requests.get")
+    @patch("backend.common.providers.crossref_provider.requests.get")
     def test__provider_does_not_call_crossref_in_test(self, mock_get):
         provider = CrossrefProvider()
         metadata, doi = provider.fetch_metadata("test_doi")
         self.assertIsNone(metadata)
         mock_get.assert_not_called()
 
-    @patch("backend.layers.thirdparty.crossref_provider.requests.get")
-    @patch("backend.layers.thirdparty.crossref_provider.CorporaConfig")
+    @patch("backend.common.providers.crossref_provider.requests.get")
+    @patch("backend.common.providers.crossref_provider.CorporaConfig")
     def test__provider_calls_crossref_if_api_key_defined(self, mock_config, mock_get):
         # Defining a mocked CorporaConfig will allow the provider to consider the `crossref_api_key`
         # not None, so it will go ahead and do the mocked call.
@@ -73,8 +73,8 @@ class TestCrossrefProvider(unittest.TestCase):
 
         self.assertDictEqual(expected_response, res)
 
-    @patch("backend.layers.thirdparty.crossref_provider.requests.get")
-    @patch("backend.layers.thirdparty.crossref_provider.CorporaConfig")
+    @patch("backend.common.providers.crossref_provider.requests.get")
+    @patch("backend.common.providers.crossref_provider.CorporaConfig")
     def test__published_doi_used_if_exists_for_preprint(self, mock_config, mock_get):
 
         # Defining a mocked CorporaConfig will allow the provider to consider the `crossref_api_key`
@@ -172,8 +172,8 @@ class TestCrossrefProvider(unittest.TestCase):
 
             self.assertDictEqual(expected_response, res)
 
-    @patch("backend.layers.thirdparty.crossref_provider.requests.get")
-    @patch("backend.layers.thirdparty.crossref_provider.CorporaConfig")
+    @patch("backend.common.providers.crossref_provider.requests.get")
+    @patch("backend.common.providers.crossref_provider.CorporaConfig")
     def test__provider_parses_authors_and_dates_correctly(self, mock_config, mock_get):
         response = Response()
         response.status_code = 200
@@ -243,8 +243,8 @@ class TestCrossrefProvider(unittest.TestCase):
 
         self.assertDictEqual(expected_response, res)
 
-    @patch("backend.layers.thirdparty.crossref_provider.requests.get")
-    @patch("backend.layers.thirdparty.crossref_provider.CorporaConfig")
+    @patch("backend.common.providers.crossref_provider.requests.get")
+    @patch("backend.common.providers.crossref_provider.CorporaConfig")
     def test__provider_unescapes_journal_correctly(self, mock_config, mock_get):
         response = Response()
         response.status_code = 200
@@ -282,8 +282,8 @@ class TestCrossrefProvider(unittest.TestCase):
 
         self.assertDictEqual(expected_response, author_data)
 
-    @patch("backend.layers.thirdparty.crossref_provider.requests.get")
-    @patch("backend.layers.thirdparty.crossref_provider.CorporaConfig")
+    @patch("backend.common.providers.crossref_provider.requests.get")
+    @patch("backend.common.providers.crossref_provider.CorporaConfig")
     def test__provider_throws_exception_if_request_fails(self, mock_config, mock_get):
         """
         Asserts a CrossrefFetchException if the GET request fails for any reason
@@ -299,8 +299,8 @@ class TestCrossrefProvider(unittest.TestCase):
         with self.assertRaises(CrossrefException):
             provider.fetch_metadata("test_doi")
 
-    @patch("backend.layers.thirdparty.crossref_provider.requests.get")
-    @patch("backend.layers.thirdparty.crossref_provider.CorporaConfig")
+    @patch("backend.common.providers.crossref_provider.requests.get")
+    @patch("backend.common.providers.crossref_provider.CorporaConfig")
     def test__provider_throws_exception_if_request_fails_with_404(self, mock_config, mock_get):
         """
         Asserts a CrossrefFetchException if the GET request fails for any reason
@@ -314,8 +314,8 @@ class TestCrossrefProvider(unittest.TestCase):
         with self.assertRaises(CrossrefDOINotFoundException):
             provider.fetch_metadata("test_doi")
 
-    @patch("backend.layers.thirdparty.crossref_provider.requests.get")
-    @patch("backend.layers.thirdparty.crossref_provider.CorporaConfig")
+    @patch("backend.common.providers.crossref_provider.requests.get")
+    @patch("backend.common.providers.crossref_provider.CorporaConfig")
     def test__provider_throws_exception_if_request_fails_with_non_2xx_code(self, mock_config, mock_get):
         """
         Asserts a CrossrefFetchException if the GET request return a 500 error (any non 2xx will work)
@@ -330,8 +330,8 @@ class TestCrossrefProvider(unittest.TestCase):
         with self.assertRaises(CrossrefFetchException):
             provider.fetch_metadata("test_doi")
 
-    @patch("backend.layers.thirdparty.crossref_provider.requests.get")
-    @patch("backend.layers.thirdparty.crossref_provider.CorporaConfig")
+    @patch("backend.common.providers.crossref_provider.requests.get")
+    @patch("backend.common.providers.crossref_provider.CorporaConfig")
     def test__provider_throws_exception_if_request_cannot_be_parsed(self, mock_config, mock_get):
         """
         Asserts an CrossrefParseException if the GET request succeeds but cannot be parsed
@@ -351,6 +351,24 @@ class TestCrossrefProvider(unittest.TestCase):
         # Make sure that the parent CrossrefException will also be caught
         with self.assertRaises(CrossrefException):
             provider.fetch_metadata("test_doi")
+
+    @patch("backend.common.providers.crossref_provider.requests.get")
+    def test__get_title_and_citation_from_doi(self, mock_get):
+        provider = CrossrefProvider()
+        mock_response = Mock()
+        mock_response.status_code = 200
+        mock_response.json.return_value = {
+            "message": {
+                "title": ["Test Title"],
+                "author": [{"family": "Doe", "given": "John"}],
+                "container-title": ["Test Journal"],
+                "created": {"date-parts": [[2022]]},
+            }
+        }
+        mock_get.return_value = mock_response
+
+        result = provider.get_title_and_citation_from_doi("10.1016/j.cell.2019.11.025")
+        self.assertEqual(result, "Test Title\n\n - Doe, John et al. (2022) Test Journal")
 
 
 if __name__ == "__main__":
