+++ tests/unit/backend/common/test_doi.py
@@ -1,0 +1,15 @@
+from backend.common.doi import clean_doi
+
+
+def test__clean_doi(self):
+    test_cases = [
+        ("10.1016/j.cell.2019.11.025", "10.1016/j.cell.2019.11.025"),
+        ("DOI: 10.1016/j.cell.2019.11.025.", "10.1016/j.cell.2019.11.025"),
+        (" DOI: 10.1016/j.cell.2019.11.025 ", "10.1016/j.cell.2019.11.025"),
+        ("10.1016/j.cell.2019.11.025. ", "10.1016/j.cell.2019.11.025"),
+        ("", ""),
+    ]
+
+    for doi, expected in test_cases:
+        with self.subTest(doi=doi):
+            self.assertEqual(clean_doi(doi), expected)
